requirements.txt
google-generativeai
langchain-google-genai
langchain-community
langchain-chroma
fastapi
streamlit
unstructured